pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "fast-food-optimizer"
version = "0.1.0"
description = "Route optimization system for breaking fast food restaurant visit world record"
readme = "README.md"
authors = [
    {name = "Jacob Sargent", email = "jacob@example.com"}
]
license = {text = "MIT"}
requires-python = ">=3.9"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]

dependencies = [
    "googlemaps>=4.10.0",
    "requests>=2.31.0",
    "pandas>=2.0.3",
    "numpy>=1.24.3",
    "scipy>=1.11.1",
    "geopy>=2.3.0",
    "shapely>=2.0.1",
    "geopandas>=0.13.2",
    "scikit-learn>=1.3.0",
    "ortools>=9.7.2996",
    "folium>=0.14.0",
    "matplotlib>=3.7.2",
    "plotly>=5.15.0",
    "click>=8.1.6",
    "python-dotenv>=1.0.0",
    "pydantic>=2.1.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "black>=23.7.0",
    "mypy>=1.4.1",
    "flake8>=6.1.0",
    "isort>=5.12.0",
]

[project.scripts]
ffro = "fast_food_optimizer.cli:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 100
target-version = ["py39", "py310", "py311"]
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
line_length = 100

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--verbose",
    "--cov=fast_food_optimizer",
    "--cov-report=term-missing",
    "--cov-report=html",
]
